requirements.txt
audioop-lts==0.2.2
paho-mqtt==2.1.0
python-dotenv==1.1.1
pyVoIP @ git+https://github.com/matmemo/pyVoIP.git@32af418fb5d36e6ad2493f0fd77d9f99db30011d
